We only need one initialization component for now.
The reason it lives outside of core is the end state can either be:
1- uber plugin module that depends on core+everything else with one initialization component handling everything. (+ this is nice to see what's going on)
or
2- initialization component in each vertical module.

We definitely don't want it in core as core should not depend on anything upstream. (verticals or uber plugin)
